Move algorithm and other relevant code into separate file (no actual code changes)
